+++ src/Slide.js
@@ -1,0 +1,26 @@
+import styled from 'styled-components'
+import { space, color } from 'styled-system'
+
+export const Slide = styled.div([], {
+  flex: 'none',
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'center',
+  flexDirection: 'column',
+  overflow: 'hidden',
+  width: '100%',
+  height: '100%',
+  '@media print': {
+    width: '100vw',
+    height: '100vh',
+    pageBreakAfter: 'always',
+    pageBreakInside: 'avoid',
+    WebkitPrintColorAdjust: 'exact'
+  }
+}, space, color)
+
+Slide.defaultProps = {
+  px: [ 4, 5, 6 ]
+}
+
+export default Slide
